fix(client): harden terminal run recovery
