+++ client/src/hooks/SSE/resumableRecovery.ts
@@ -11,8 +11,10 @@ export type DisconnectedRunRecovery = {
 const terminalEventQueryKey = (conversationId: string) =>
   ['resumable-terminal-event', conversationId] as const;
 
+const disconnectedRunQueryRoot = ['resumable-disconnected-run'] as const;
+
 const disconnectedRunQueryKey = (conversationId: string) =>
-  ['resumable-disconnected-run', conversationId] as const;
+  [...disconnectedRunQueryRoot, conversationId] as const;
 
 export function markTerminalEventSeen(queryClient: QueryClient, conversationId: string) {
   queryClient.setQueryData(terminalEventQueryKey(conversationId), true);
@@ -43,6 +45,9 @@ export function setDisconnectedRunRecovery(
   conversationId: string,
   recovery: DisconnectedRunRecovery,
 ) {
+  // This is session state, not fetched server data. It must survive the
+  // default five-minute React Query GC while a disconnected job keeps running.
+  queryClient.setQueryDefaults(disconnectedRunQueryRoot, { cacheTime: Infinity });
   queryClient.setQueryData(disconnectedRunQueryKey(conversationId), recovery);
 }
